+++ src/upnp/cont_dir_service.cc
@@ -123,7 +123,7 @@ void ContentDirectoryService::doBrowse(ActionRequest& request)
     auto upnpClass = parent->getClass();
     if (sortCriteria.empty() && (startswith(upnpClass, UPNP_CLASS_MUSIC_ALBUM) || startswith(upnpClass, UPNP_CLASS_PLAYLIST_CONTAINER)))
         flag |= BROWSE_TRACK_SORT;
-    else if (quirks->hasFlag(QUIRK_FLAG_FORCE_SORT_CRITERIA_TITLE))
+    else if (quirks->hasFlag(Quirk::ForceSortCriteriaTitle))
         sortCriteria = fmt::format("+{}", MetaEnumMapper::getMetaFieldName(MetadataFields::M_TITLE));
     if (filter.empty())
         filter = "*";
@@ -132,7 +132,7 @@ void ContentDirectoryService::doBrowse(ActionRequest& request)
 
     auto param = BrowseParam(parent, flag);
 
-    param.setDynamicContainers(!quirks || !quirks->checkFlags(QUIRK_FLAG_SAMSUNG_HIDE_DYNAMIC));
+    param.setDynamicContainers(!quirks || !quirks->hasFlag(Quirk::SamsungHideDynamic));
     param.setStartingIndex(startingIndex);
     param.setRequestedCount(requestedCount);
     param.setSortCriteria(trimString(sortCriteria));
@@ -231,9 +231,9 @@ void ContentDirectoryService::doSearch(ActionRequest& request)
     didlLiteRoot.append_attribute(UPNP_XML_DC_NAMESPACE_ATTR) = UPNP_XML_DC_NAMESPACE;
     didlLiteRoot.append_attribute(UPNP_XML_UPNP_NAMESPACE_ATTR) = UPNP_XML_UPNP_NAMESPACE;
     didlLiteRoot.append_attribute(UPNP_XML_SEC_NAMESPACE_ATTR) = UPNP_XML_SEC_NAMESPACE;
-    if (quirks->checkFlags(QUIRK_FLAG_PV_SUBTITLES))
+    if (quirks->hasFlag(Quirk::PvSubtitles))
         didlLiteRoot.append_attribute("xmlns:pv") = "http://www.pv.com/pvns/";
-    if (sortCriteria.empty() || quirks->hasFlag(QUIRK_FLAG_FORCE_SORT_CRITERIA_TITLE)) {
+    if (sortCriteria.empty() || quirks->hasFlag(Quirk::ForceSortCriteriaTitle)) {
         sortCriteria = fmt::format("+{}", MetaEnumMapper::getMetaFieldName(MetadataFields::M_TITLE));
     }
     if (filter.empty())